+++ inseq/attr/feat/ops/__init__.py
@@ -2,10 +2,12 @@ from .discretized_integrated_gradients import DiscretetizedIntegratedGradients
 from .lime import Lime
 from .monotonic_path_builder import MonotonicPathBuilder
 from .sequential_integrated_gradients import SequentialIntegratedGradients
+from .value_zeroing import ValueZeroing
 
 __all__ = [
     "DiscretetizedIntegratedGradients",
     "MonotonicPathBuilder",
+    "ValueZeroing",
     "Lime",
     "SequentialIntegratedGradients",
 ]