+++ inseq/utils/torch_utils.py
@@ -5,10 +5,13 @@ from typing import TYPE_CHECKING, Callable, Literal, Optional, Union
 import torch
 import torch.nn.functional as F
 from jaxtyping import Int, Num
+from torch import nn
 from torch.backends.cuda import is_built as is_cuda_built
 from torch.backends.mps import is_available as is_mps_available
 from torch.backends.mps import is_built as is_mps_built
 from torch.cuda import is_available as is_cuda_available
+
+from .typing import OneOrMoreIndices
 
 if TYPE_CHECKING:
     pass
@@ -244,3 +247,118 @@ def get_default_device() -> str:
         return "cpu"
     else:
         return "cpu"
+
+
+def find_block_stack(module):
+    """Recursively searches for the first instance of a `nn.ModuleList` submodule within a given `torch.nn.Module`.
+
+    Args:
+        module (:obj:`torch.nn.Module`): A Pytorch :obj:`nn.Module` object.
+
+    Returns:
+        :obj:`torch.nn.ModuleList`: The first instance of a :obj:`nn.Module` submodule found within the given object.
+        None: If no `nn.ModuleList` submodule is found within the given `nn.Module` object.
+    """
+    # Check if the current module is an instance of nn.ModuleList
+    if isinstance(module, nn.ModuleList):
+        return module
+
+    # Recursively search for nn.ModuleList in the submodules of the current module
+    for submodule in module.children():
+        module_list = find_block_stack(submodule)
+        if module_list is not None:
+            return module_list
+
+    # If nn.ModuleList is not found in any submodules, return None
+    return None
+
+
+def validate_indices(
+    scores: torch.Tensor,
+    dim: int = -1,
+    indices: Optional[OneOrMoreIndices] = None,
+) -> OneOrMoreIndices:
+    """Validates a set of indices for a given dimension of a tensor of scores. Supports single indices, spans and lists
+    of indices, including negative indices to specify positions relative to the end of the tensor.
+
+    Args:
+        scores (torch.Tensor): The tensor of scores.
+        dim (int, optional): The dimension of the tensor that will be indexed. Defaults to -1.
+        indices (Union[int, tuple[int, int], list[int], None], optional):
+            - If an integer, it is interpreted as a single index for the dimension.
+            - If a tuple of two integers, it is interpreted as a span of indices for the dimension.
+            - If a list of integers, it is interpreted as a list of individual indices for the dimension.
+
+    Returns:
+        ``Union[int, tuple[int, int], list[int]]``: The validated list of positive indices for indexing the dimension.
+    """
+    if dim >= scores.ndim:
+        raise IndexError(f"Dimension {dim} is greater than tensor dimension {scores.ndim}")
+    n_units = scores.shape[dim]
+    if not isinstance(indices, (int, tuple, list)) and indices is not None:
+        raise TypeError(
+            "Indices must be an integer, a (start, end) tuple of indices representing a span, a list of individual"
+            " indices or a single index."
+        )
+    if hasattr(indices, "__iter__"):
+        if len(indices) == 0:
+            raise RuntimeError("An empty sequence of indices is not allowed.")
+        if len(indices) == 1:
+            indices = indices[0]
+
+    if isinstance(indices, int):
+        if indices not in range(-n_units, n_units):
+            raise IndexError(f"Index out of range. Scores only have {n_units} units.")
+        indices = indices if indices >= 0 else n_units + indices
+        return torch.tensor(indices)
+    else:
+        if indices is None:
+            indices = (0, n_units)
+            logger.info("No indices specified. Using all indices by default.")
+
+        # Convert negative indices to positive indices
+        if hasattr(indices, "__iter__"):
+            indices = type(indices)([h_idx if h_idx >= 0 else n_units + h_idx for h_idx in indices])
+        if not hasattr(indices, "__iter__") or (
+            len(indices) == 2 and isinstance(indices, tuple) and indices[0] >= indices[1]
+        ):
+            raise RuntimeError(
+                "A (start, end) tuple of indices representing a span, a list of individual indices"
+                " or a single index must be specified."
+            )
+        max_idx_val = n_units if isinstance(indices, list) else n_units + 1
+        if not all(h in range(-n_units, max_idx_val) for h in indices):
+            raise IndexError(f"One or more index out of range. Scores only have {n_units} units.")
+        if len(set(indices)) != len(indices):
+            raise IndexError("Duplicate indices are not allowed.")
+        if isinstance(indices, tuple):
+            return torch.arange(indices[0], indices[1])
+        else:
+            return torch.tensor(indices)
+
+
+def pad_with_nan(t: torch.Tensor, dim: int, pad_size: int, front: bool = False) -> torch.Tensor:
+    """Utility to pad a tensor with nan values along a given dimension."""
+    nan_tensor = torch.ones(
+        *t.shape[:dim],
+        pad_size,
+        *t.shape[dim + 1 :],
+        device=t.device,
+    ) * float("nan")
+    if front:
+        return torch.cat([nan_tensor, t], dim=dim)
+    return torch.cat([t, nan_tensor], dim=dim)
+
+
+def recursive_get_submodule(parent: nn.Module, target: str) -> Optional[nn.Module]:
+    if target == "":
+        return parent
+    mod = None
+    if hasattr(parent, target):
+        mod = getattr(parent, target)
+    else:
+        for submodule in parent.children():
+            mod = recursive_get_submodule(submodule, target)
+            if mod is not None:
+                break
+    return mod